+++ lib/src/executor_io/executor_leader_io.dart
@@ -73,7 +73,8 @@ class ExecutorLeader extends Executor {
   }
 
   @override
-  Iterable<TaskWrapper<dynamic>> get unfinishedTasks => _tasks;
+  Iterable<ConcurrentTask<dynamic>> get unfinishedTasks =>
+      _tasks.map((e) => e.task);
 
   /// process message for master
   void _messageProcessor(dynamic message) {